% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary.R
\name{fecundity}
\alias{fecundity}
\title{Fecundity at weight function with base parameters from Tanasichuk et al. (1993)}
\usage{
fecundity(w, F1 = exp(4.69), F2 = 1.13)
}
\description{
Fecundity at weight function with base parameters from Tanasichuk et al. (1993)
}
