% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assessment.R
\name{kalman_assess}
\alias{kalman_assess}
\title{Apply a Kalman filter to a time series}
\usage{
kalman_assess(TS, sd1, sd2, fit = FALSE, full.ts = FALSE, log = TRUE)
}
\description{
Apply a Kalman filter to a time series
}

