% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary.R
\name{linear_odes}
\alias{linear_odes}
\title{Define a system of ordinary differential equations from a matrix}
\usage{
linear_odes(t, state, A)
}
\description{
Define a system of ordinary differential equations from a matrix
}
