% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stochasticity.R
\name{ran_stray_prob}
\alias{ran_stray_prob}
\title{Generate a series of random stray matrices using a mean stray probability matrix}
\usage{
ran_stray_prob(stray_mat, n_iter, scale)
}
\arguments{
\item{stray_mat}{a mean stray matrix.}

\item{n_iter}{number of observations.}

\item{scale}{higher values provide lower variability in random samples.}
}
\description{
Generate a series of random stray matrices by sampling from the Dirichlet distribution with a mean stray probability matrix and scale parameter
}

