% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stochasticity.R
\name{spat_temp_ts}
\alias{spat_temp_ts}
\title{Generate a spatially and temporally correlated time series}
\usage{
spat_temp_ts(n_iter, n_loc, site_sd, spat_sd, phi, cor_mat = NULL,
  log = TRUE)
}
\arguments{
\item{n_iter}{number of observations in the multivariate time series.}

\item{n_loc}{number of locatons.}

\item{site_sd}{univariate standard devation of each time series.}

\item{spat_sd}{scale of spatial correlation.}

\item{phi}{AR(1) model coefficient.}

\item{cor_mat}{user supplied correlation matrix (if unspecified will be generated from the spat_sd parameter).}
}
\description{
Simulates a multivarate normal, AR(1) time series that can be used for random errors.  Each time point includes vector of correlated deviations, which exhibit either Gaussian spatial correlations OR a specified correlation matrix.   If log = TRUE, the sd is adjusted such that the final expected sd is equal to that specified.

Step 1: Generate or supply a spatial correlation matrix (cor_mat)

Step 2: Generate a variance-covariance matrix using site_sd and cor_mat

Step 3: Create a time series of multivariate normal vectors (each vector is of length n_loc) with n_iter vector observations in the time series.

Step 4: Run each of the n_log time series through an AR(1) filter with first order correlation phi

Step 5: Generate observed correlation matrix and vector of first order correlations the time series
}
\examples{

### generate the time series
errors <- spat_temp_ts(n_iter = 100, n_loc = 5, 
                       site_sd = 0.5, spat_sd = 2, 
                       phi = 0.5)

# examine correlation matrix of time series
errors$cor_mat

# examine pacf of each time series
errors$pacf

# examine sd of each time series
apply(errors$ts,2,sd)

# plot the time series
matplot(errors$ts,type= "l")
}
\seealso{
\code{\link{spat_cor_mat}}
}

