% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary.R
\name{weight}
\alias{weight}
\title{Power length-weight relationship (from Martell et al. 2012)}
\usage{
weight(length, a = 4.5e-06, b = 3.127)
}
\description{
Power length-weight relationship (from Martell et al. 2012)
}
