% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stochasticity.R
\name{spat_cor_mat}
\alias{spat_cor_mat}
\title{Generate a spatially correlated matrix}
\usage{
spat_cor_mat(n_loc, spat_sd = 1, spat_scale = NULL, sumto1 = FALSE)
}
\arguments{
\item{n_loc}{number of locatons.}

\item{spat_sd}{scale of spatial correlation for the Gaussian case.}

\item{sumto1}{if TRUE makes the columns sum to 1.}

\item{spat_sd}{scale of spatial correlation for the Cauchy case.}
}
\description{
Generate a spatially correlated matrix where correlations decrease in space via a Gaussian process (with spat_sd) OR a Cauchy process (with spat_scale)
}
\examples{
par(mfrow = c(3,2),mai= c(0.75,0.75,0.75,0.75),mgp = c(1.5,0.5,0.1))

matplot(spat_cor_mat(n_loc=10, spat_scale = 1),type= "l",ylab= "correlation",xlab= "site",main= "Cauchy, scale= 1")
matplot(spat_cor_mat(n_loc=10, spat_sd = 1),type= "l",ylab= "correlation",xlab= "site",main= "Gaussian, sd= 1")
matplot(spat_cor_mat(n_loc=10, spat_scale = 2),type= "l",ylab= "correlation",xlab= "site",main= "Cauchy, scale= 2")
matplot(spat_cor_mat(n_loc=10, spat_sd = 2),type= "l",ylab= "correlation",xlab= "site",main= "Gaussian, sd= 2")
matplot(spat_cor_mat(n_loc=10, spat_scale = 10),type= "l",ylab= "correlation",xlab= "site",main= "Cauchy, scale= 10")
matplot(spat_cor_mat(n_loc=10, spat_sd = 10),type= "l",ylab= "correlation",xlab= "site",main= "Gaussian, sd= 10")
}

